backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# Database drivers
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
motor==3.3.2  # MongoDB async driver
pymongo==4.6.0
redis==5.0.1
elasticsearch==8.11.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Data Science & ML
pandas==2.1.3
numpy==1.24.4
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
tensorflow==2.15.0
torch==2.1.1
transformers==4.36.0
prophet==1.1.5
statsmodels==0.14.0
plotly==5.17.0

# NLP & AI
langchain==0.0.340
langchain-openai==0.0.2
openai==1.3.7
sentence-transformers==2.2.2
spacy==3.7.2

# Geospatial
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1
folium==0.15.1
geopy==2.4.1

# HTTP Client & APIs
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Background Tasks
celery==5.3.4
flower==2.0.1

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# File processing
openpyxl==3.1.2
xlsxwriter==3.1.9
reportlab==4.0.7
python-docx==1.1.0

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Image processing
Pillow==10.1.0

# Notification services
twilio==8.10.3
sendgrid==6.11.0

# Data validation
cerberus==1.3.5
marshmallow==3.20.2

# WebSocket support
websockets==12.0

# GraphQL
graphene==3.3
strawberry-graphql[fastapi]==0.215.1

# Caching
aiocache==0.12.2

# File storage
boto3==1.34.0  # AWS S3